move main file and plugin to a cmd dir
